requirements.txt
torchaudio==2.2.1
torchvision==0.17.1
torchao==0.1
torchtune==0.1.1
setuptools==69.5.1
yacs==0.1.8
tensorboardX==2.6.2.2
tensorboard==2.16.2
moviepy==1.0.3
colorama==0.4.6
einops==0.8.0
tqdm==4.66.4
opencv-python-headless==4.9.0.80
gymnasium[atari,accept-rom-license]==0.29.1
line_profiler==4.1.3
mamba-ssm==1.2.0.post1
kornia==0.7.3
causal-conv1d==1.2.0.post2
wandb
pytorch_warmup==0.1.1
numpy==1.24.3
pandas==2.2.2
memory-maze==1.0.3
gym